requirements.txt
opencv-python-headless
pillow
streamlit-camera-input-live